Add files via upload
Task 1
Write a program that does the following:a) Stores a random number (1-10) in a variable –see hint below.b) Asks a user for their name and stores this in a variable.c) Asks a user to guess the number between 1 and 10.d) Tells the user whetherthey have guessed correctly.

Task 2
Write a program that asks a user for their favourite number between 1 and 100 and then tells them a joke based on the number. You should use a minimum of 3 jokes.

Task 3
Write a program that allows user to enter their favourite starter, main course, dessert and drink.Concatenate these and output a message which says –“Your favourite meal is  .........with a glass of....”

Task 4
A motorbike costs £2000 and loses 10% of its value every year. Using a loop, print the value of the bike every following year until it falls below £1000.

Task 5
Write a program which will ask for two numbers from a user. Then offer a menu to the user giving them a choice of operator:e.g. –Enter “a” if you want to add“b” if you want to subtract, Include +, -, /, *, ** square (to the power of). Once the user has selected which operator they wish to use, perform the calculation.